fixed rename and cleaned base class
