+++ test/library/calibration/test_rabi.py
@@ -28,7 +28,7 @@ from qiskit_experiments.library import Rabi, EFRabi
 from qiskit_experiments.curve_analysis.standard_analysis.oscillation import OscillationAnalysis
 from qiskit_experiments.data_processing.data_processor import DataProcessor
 from qiskit_experiments.data_processing.nodes import Probability
-from qiskit_experiments.test.iq_pulse_backend import SingleTransmonTestBackend
+from qiskit_experiments.test.pulse_backend import SingleTransmonTestBackend
 from qiskit_experiments.framework.experiment_data import ExperimentStatus
