+++ test/library/calibration/test_rough_frequency.py
@@ -19,7 +19,7 @@ from qiskit_experiments.framework import BackendData
 from qiskit_experiments.library import RoughFrequencyCal
 from qiskit_experiments.calibration_management import Calibrations
 from qiskit_experiments.calibration_management.basis_gate_library import FixedFrequencyTransmon
-from qiskit_experiments.test.iq_pulse_backend import SingleTransmonTestBackend
+from qiskit_experiments.test.pulse_backend import SingleTransmonTestBackend
 
 
 class TestRoughFrequency(QiskitExperimentsTestCase):
